Finish implementing MultiThreadedCache, and add stress test.
This cache has no locks on access, and only has contention on a cache
miss. It only ever holds a shared lock for a constant-time duration,
never while executing user code. A Task requesting a key that is
already being computed on another Task will block while that computation
is being performed. By taking advantage of the append-only properties of
a cache, the cache can be duplicated per-Thread, to avoid locking in the
common case.