Add benchmark test measuring parallel scaling.
The results are a bit lukewarm (glad we measured it), but look alright.

Here's a few runs, each with different JULIA_NUM_THREADS:
```julia
┌ Info: benchmark results
│   Threads.nthreads() = 1
│   time_serial = 0.013336288
│   time_parallel = 0.09071632
└   time_baseline = 0.115363526
```
```julia
┌ Info: benchmark results
│   Threads.nthreads() = 2
│   time_serial = 0.011262138
│   time_parallel = 0.097021203
└   time_baseline = 0.139031655
```
```julia
┌ Info: benchmark results
│   Threads.nthreads() = 20
│   time_serial = 0.011997677
│   time_parallel = 0.658225544
└   time_baseline = 1.032283809
```
```julia
┌ Info: benchmark results
│   Threads.nthreads() = 100
│   time_serial = 0.013902211
│   time_parallel = 1.999772731
└   time_baseline = 9.314424419
```

So it definitely does not scale as well as a single-threaded codebase
would have scaled. But it also definitely scales better than the
baseline, which is a Mutex around a Dict.

Maybe there will be some places to improve our contention in the future :)